requirements.txt
# ============================================
# 한국 주식 AI 스윙 트레이딩 시스템
# 패키지 의존성 목록
# ============================================

# ===== 핵심 라이브러리 =====
python-dotenv==1.0.0       # 환경 변수 관리
pydantic==2.5.3            # 데이터 검증
pydantic-settings==2.1.0   # 설정 관리

# ===== 데이터 처리 =====
pandas==2.1.4              # 데이터 분석
numpy==1.26.3              # 수치 계산
scipy==1.11.4              # 통계 분석

# ===== HTTP 클라이언트 =====
httpx==0.25.2              # 비동기 HTTP 클라이언트
requests==2.31.0           # 동기 HTTP 클라이언트

# ===== 웹소켓 =====
websockets==12.0           # 실시간 시세 수신

# ===== 웹 크롤링 =====
beautifulsoup4==4.12.2     # HTML 파싱
lxml==4.9.4                # XML/HTML 파서

# ===== AI API =====
anthropic==0.18.1          # Claude API 클라이언트

# ===== 스케줄링 =====
apscheduler==3.10.4        # 작업 스케줄러

# ===== 로깅 =====
loguru==0.7.2              # 구조화된 로깅

# ===== 데이터베이스 =====
# SQLite는 Python 내장 모듈이므로 별도 설치 불필요
# SQLAlchemy==2.0.25       # ORM (선택사항)

# ===== 개발 도구 (선택사항) =====
# black==23.12.1           # 코드 포맷터
# ruff==0.1.9              # 린터
# pytest==7.4.4            # 테스트 프레임워크
# mypy==1.8.0              # 타입 체커

# ===== 시각화 (선택사항) =====
# plotly==5.18.0           # 차트 시각화
